+++ witp/src/main/java/dev/efnilite/ip/api/event/ParkourBlockGenerateEvent.java
@@ -5,17 +5,19 @@ import dev.efnilite.ip.player.ParkourPlayer;
 import dev.efnilite.vilib.event.EventWrapper;
 import org.bukkit.block.Block;
 
+import java.util.List;
+
 /**
- * Gets called when a block is generated. Read-only.
+ * Gets called when a new jump is generated. Read-only.
  */
 public class ParkourBlockGenerateEvent extends EventWrapper {
 
-    public final Block block;
+    public final List<Block> blocks;
     public final ParkourGenerator generator;
     public final ParkourPlayer player;
 
-    public ParkourBlockGenerateEvent(Block block, ParkourGenerator generator, ParkourPlayer player) {
-        this.block = block;
+    public ParkourBlockGenerateEvent(List<Block> blocks, ParkourGenerator generator, ParkourPlayer player) {
+        this.blocks = blocks;
         this.generator = generator;
         this.player = player;
     }