+++ witp/src/main/java/dev/efnilite/ip/menu/community/SingleLeaderboardMenu.java
@@ -40,7 +40,7 @@ public class SingleLeaderboardMenu {
 
         // init vars
         ParkourUser user = ParkourUser.getUser(player);
-        String locale = user == null ? Option.OPTIONS_DEFAULTS.get(ParkourOption.LANG) : user.getLocale();
+        String locale = user == null ? Option.OPTIONS_DEFAULTS.get(ParkourOption.LANG) : user.locale;
         PagedMenu menu = new PagedMenu(3, Locales.getString(player, "%s.name".formatted(ParkourOption.LEADERBOARDS.path)));
 
         List<MenuItem> items = new ArrayList<>();