+++ witp/src/main/java/dev/efnilite/ip/menu/play/SingleMenu.java
@@ -28,7 +28,7 @@ public class SingleMenu {
 
     public void open(Player player) {
         ParkourUser user = ParkourUser.getUser(player);
-        String locale = user == null ? Option.OPTIONS_DEFAULTS.get(ParkourOption.LANG) : user.getLocale();
+        String locale = user == null ? Option.OPTIONS_DEFAULTS.get(ParkourOption.LANG) : user.locale;
 
         List<Mode> modes = IP.getRegistry().getModes();